Pass render locals/opts to template engine
The current Hobbit::Render#render method accepts arguments for locals
and options, but doesn't pass these down to the actual private _render
call. In the case of locals, the value is lost entirely and never used
anywhere. This modifies the calls to _render to forward those values to
the templater.